MVC_sistema_leitura_hidrometros/MVC_sistema_leitura_hidrometros/requirements.txt
flask==3.0.3
flask_sqlalchemy==3.1.1
flask-socketio==5.3.6
eventlet==0.36.1
pymysql==1.1.0
paho-mqtt==1.6.1
python-dotenv==1.0.1
flask-migrate==4.0.7
PyJWT==2.9.0
# Decimal precision
